// frontend/src/api/model/organizationDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { AddressDto } from './addressDto';
import type { OrganizationDtoType } from './organizationDtoType';

/**
 * объект
 */
export interface OrganizationDto {
  id: number;
  name?: string;
  officialAddress?: AddressDto;
  annualTurnover?: number;
  employeesCount?: number;
  fullName?: string;
  rating?: number;
  type?: OrganizationDtoType;
}