// frontend/src/api/model/organizationDtoType.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export type OrganizationDtoType = typeof OrganizationDtoType[keyof typeof OrganizationDtoType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const OrganizationDtoType = {
  COMMERCIAL: 'COMMERCIAL',
  PUBLIC: 'PUBLIC',
  TRUST: 'TRUST',
  OPEN_JOINT_STOCK_COMPANY: 'OPEN_JOINT_STOCK_COMPANY',
} as const;